src/test/java/practice/array/RockPaperScissorsTest.java
package practice.array;

import static org.assertj.core.api.Assertions.assertThat;
import java.util.stream.Stream;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

/**
 * <h1>3. 가위 바위 보</h1>
 * <h1>설명</h1>
 *
 * A, B 두 사람이 가위바위보 게임을 합니다. 총 N번의 게임을 하여 A가 이기면 A를 출력하고, B가 이기면 B를 출력합니다. 비길 경우에는 D를 출력합니다.
 *
 * 가위, 바위, 보의 정보는 1:가위, 2:바위, 3:보로 정하겠습니다.
 *
 * 두 사람의 각 회의 가위, 바위, 보 정보가 주어지면 각 회를 누가 이겼는지 출력하는 프로그램을 작성하세요.
 *
 * <h1>입력</h1>
 * 첫 번째 줄에 게임 횟수인 자연수 N(1<=N<=100)이 주어집니다.
 *
 * 두 번째 줄에는 A가 낸 가위, 바위, 보 정보가 N개 주어집니다.
 *
 * 세 번째 줄에는 B가 낸 가위, 바위, 보 정보가 N개 주어집니다.
 *
 * <h1>출력</h1>
 * 각 줄에 각 회의 승자를 출력합니다. 비겼을 경우는 D를 출력합니다.
 */
@DisplayName("배열 - 가위 바위 보")
class RockPaperScissorsTest {

    private RockPaperScissors solve = new RockPaperScissors();

    @MethodSource
    @ParameterizedTest
    void solution(final int a, final int b, final String expected) {
        final String actual = solve.solution(a, b);
        assertThat(actual).isEqualTo(expected);
    }

    private static Stream<Arguments> solution() {
        return Stream.of(
            Arguments.of(2, 1, "A"),
            Arguments.of(3, 1, "B"),
            Arguments.of(3, 2, "A"),
            Arguments.of(1, 2, "B"),
            Arguments.of(3, 3, "D")
        );
    }

}
